+++ docs/changelog.md
@@ -1,8 +1,11 @@
+# 0.0.27
+- Navidrome works!
+- Tidy up request/response structure in backend
+- Tidy Settings page
+
 # 0.0.26
 - Make email required
 - Add basic navidrome/subsonic connection
-- Tidy up request/response structure in backend
-- Tidy Settings page
 
 # 0.0.25
 - Images now pull from spotify if setup!